projects/requirements.txt
pandas==1.4.3
numpy==1.21.4
dagster
pandas
dagit
openpyxl
minio
unidecode
psycopg2-binary